requirement.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
django-filter==23.4
djangorestframework==3.14.0
Markdown==3.5.1
Pillow==10.1.0
protobuf==4.21.12
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
